Refactor Product issues and fix issues when getting stats for a fresh MC account

+++ src/DB/Table/MerchantIssueTable.php
@@ -68,6 +68,10 @@ SQL;
 	 * @param string $source       The source of the issues. Default is 'mc'.
 	 */
 	public function delete_specific_product_issues( array $products_ids, string $source = 'mc' ): void {
+		if ( empty( $products_ids ) ) {
+			return;
+		}
+
 		$placeholder = '(' . implode( ',', array_fill( 0, count( $products_ids ), '%d' ) ) . ')';
 		$this->wpdb->query( $this->wpdb->prepare( "DELETE FROM `{$this->get_sql_safe_name()}` WHERE `product_id` IN {$placeholder} AND `source` = %s", array_merge( $products_ids, [ $source ] ) ) ); // phpcs:ignore WordPress.DB.PreparedSQL
 	}
